requirements.txt
streamlit
pillow
pytesseract
pypdf2
openai
transformers
torch
